Move 'has-class' to private folder
